feat: Allow to manage multiple servers and support Bevy 0.15.x

+++ src/extension/vscode/entities/entitiesController.ts
@@ -1,40 +1,36 @@
 import * as vscode from 'vscode';
-import type { EntitySelection } from '../EntitySelection';
+import type { EntityUpdated } from '../components/components';
 import { PollingService } from '../polling';
 import { EntityTreeDataProvider } from './entitiesDataProvider';
 import type { EntityNode, EntityTreeRepository } from './entityTree';
 
 export class TreeController {
+  private repository: EntityTreeRepository;
   private treeDataProvider: EntityTreeDataProvider;
   private treeView: vscode.TreeView<EntityNode>;
   private pollingService: PollingService = new PollingService();
-  private readonly entitySelectionEmitter = new vscode.EventEmitter<EntitySelection>();
-  public readonly onSelectionChanged = this.entitySelectionEmitter.event;
-  private readonly repository: EntityTreeRepository;
+  private readonly entityNodeEmitter = new vscode.EventEmitter<EntityNode>();
+  public readonly onSelectionChanged = this.entityNodeEmitter.event;
 
   constructor(context: vscode.ExtensionContext, repository: EntityTreeRepository) {
     this.repository = repository;
-    this.treeDataProvider = new EntityTreeDataProvider(repository);
-
+    this.treeDataProvider = new EntityTreeDataProvider();
     this.treeView = vscode.window.createTreeView('bevyInspector.entities', {
       treeDataProvider: this.treeDataProvider,
       showCollapseAll: true,
       canSelectMany: false,
     });
-
+    this.repository.tree().then(this.treeDataProvider.setEntities.bind(this.treeDataProvider));
     this.treeView.onDidChangeSelection(this.handleSelectionChange.bind(this));
-    context.subscriptions.push(this.treeView);
-
     context.subscriptions.push(
-      vscode.commands.registerCommand('bevyInspector.refresh', () => {
-        // this.repository.invalidateCache();
-        this.refresh();
-      }),
+      vscode.commands.registerCommand('bevyInspector.refresh', () => this.refresh()),
       vscode.commands.registerCommand('bevyInspector.enablePolling', () => this.pollingService.enablePolling()),
       vscode.commands.registerCommand('bevyInspector.disablePolling', () => this.pollingService.disablePolling()),
       vscode.commands.registerCommand('bevyInspector.destroyEntity', async (entity) => {
-        await this.repository.destroy(entity);
-        this.refresh();
+        if (entity !== undefined && this.repository !== null) {
+          await this.repository.destroy(entity);
+          this.refresh();
+        }
       }),
     );
     vscode.workspace.onDidChangeConfiguration((e) => {
@@ -45,17 +41,19 @@ export class TreeController {
   }
 
   public refresh() {
-    this.treeDataProvider.refresh();
+    this.repository.tree().then(this.treeDataProvider.setEntities.bind(this.treeDataProvider));
     if (this.treeView.selection.length > 0) {
-      const selectedEntity = this.treeView.selection[0] as EntityNode;
-      this.entitySelectionEmitter.fire({ entityId: selectedEntity.id });
+      this.entityNodeEmitter.fire(this.treeView.selection[0]);
+    }
+  }
+
+  public refreshNames(entityUpdated: EntityUpdated) {
+    if (this.repository.isNameType(entityUpdated.typePath) && typeof entityUpdated.newValue === 'string') {
+      this.treeDataProvider.setEntityName(entityUpdated.entityId, entityUpdated.newValue);
     }
   }
 
   private handleSelectionChange(selectionChanged: vscode.TreeViewSelectionChangeEvent<EntityNode>) {
-    if (selectionChanged.selection[0]) {
-      const selectedEntity = selectionChanged.selection[0] as EntityNode;
-      this.entitySelectionEmitter.fire({ entityId: selectedEntity.id });
-    }
+    this.entityNodeEmitter.fire(selectionChanged.selection[0]);
   }
 }
